Refactor request builder.
